fix: move vcr filtering to top-level conftest
